+++ apps/ai-dial-admin/src/components/Deployments/Modals/ContainerDuplicate.tsx
@@ -6,8 +6,8 @@ import { Container } from '@/src/models/deployments/containers';
 import { useSaveValidationContext } from '@/src/context/SaveValidationContext';
 import { useI18n } from '@/src/locales/client';
 
-import IdControl from '@/src/components/EntityMainProperties/BaseProperties/Id';
-import DisplayNameControl from '@/src/components/EntityMainProperties/BaseProperties/DisplayName';
+import IdControl from '@/src/components/BaseControls/Id';
+import DisplayNameControl from '@/src/components/BaseControls/DisplayName';
 
 interface Props {
   title: string;
