+++ apps/ai-dial-admin/src/components/ModelView/ModelProperties/ModelProperties.tsx
@@ -5,7 +5,7 @@ import Defaults from '@/src/components/Defaults/Defaults';
 import UpstreamEndpoints from '@/src/components/UpstreamEndpoints/UpstreamEndpoints';
 import DeploymentProperties from '@/src/components/EntityMainProperties/Properties/DeploymentProperties';
 import ForwardAuthTokenField from '@/src/components/EntityMainProperties/ForwardAuthToken/ForwardAuthTokenField';
-import MaxRetryAttempts from '@/src/components/EntityMainProperties/BaseProperties/MaxRetryAttempts';
+import MaxRetryAttempts from '@/src/components/BaseControls/MaxRetryAttempts';
 import Limits from '@/src/components/ModelView/Limits/Limits';
 import Pricing from '@/src/components/ModelView/Pricing/Pricing';
 import TokenizerModelSwitch from '@/src/components/Models/View/TokenizerModel/Tokenizer';
