+++ apps/ai-dial-admin/src/components/TestSuites/Modals/Create/Properties.tsx
@@ -2,8 +2,8 @@
 
 import { FC } from 'react';
 
-import DescriptionControl from '@/src/components/EntityMainProperties/BaseProperties/Description';
-import DisplayNameControl from '@/src/components/EntityMainProperties/BaseProperties/DisplayName';
+import DescriptionControl from '@/src/components/BaseControls/Description';
+import DisplayNameControl from '@/src/components/BaseControls/DisplayName';
 import { TestSuite } from '@/src/models/evaluation/test-suite';
 
 interface Props {
