+++ apps/ai-dial-admin/src/components/UpstreamEndpoints/Endpoint/Endpoint.tsx
@@ -17,7 +17,7 @@ import { useIsTabletScreen } from '@/src/hooks/use-is-tablet-screen';
 import { useI18n } from '@/src/locales/client';
 import { DialEndpointExtraData, DialModelEndpoint } from '@/src/models/dial/model';
 import { isDangerEndpoint } from '@/src/utils/validation/url-error';
-import EndpointControl from '@/src/components/EntityMainProperties/BaseProperties/Endpoint/Endpoint';
+import EndpointControl from '@/src/components/BaseControls/Endpoint/Endpoint';
 import ExtraDataField from '../ExtraData/ExtraDataField';
 import WarningIcon from './WarningIcon';
